requirements.txt
scikit-learn
numpy
matplotlib
umap-learn
pandas
transformers
torch